+++ http/src/main/scala/com/twitter/finatra/http/Tls.scala
@@ -8,7 +8,7 @@ import java.io.File
 /**
  * A helper trait for serving HTTPS requests with standard TLS. To use,
  * mix into an [[HttpServer]] and pass the `cert.path` and `key.path` flags
- * or set the defaults accordingly.
+ * and optionally `chain.path` flag if using a custom chain or set the defaults accordingly.
  *
  * {{{
  *   object MyServiceMain extends MyService
@@ -39,11 +39,24 @@ trait Tls { self: HttpServer =>
   private val keyPathFlag =
     flag("key.path", defaultKeyPath, "path to SSL key")
 
+  protected def defaultChainPath: String = ""
+
+  private val chainPathFlag =
+    flag("chain.path", defaultChainPath, "path to SSL chain")
+
   override protected def configureHttpsServer(server: Http.Server): Http.Server = {
+    val chainPath = chainPathFlag()
     val keyCredentials =
-      KeyCredentials.CertAndKey(
-        certificateFile = new File(certificatePathFlag()),
-        keyFile = new File(keyPathFlag()))
+      if (chainPath.isEmpty) {
+        KeyCredentials.CertAndKey(
+          certificateFile = new File(certificatePathFlag()),
+          keyFile = new File(keyPathFlag()))
+      } else {
+        KeyCredentials.CertKeyAndChain(
+          certificateFile = new File(certificatePathFlag()),
+          keyFile = new File(keyPathFlag()),
+          caCertificateFile = new File(chainPath))
+      }
 
     server.withTransport
       .tls(SslServerConfiguration(keyCredentials))